# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"Api example for alexis","description":"Spring.boot-api-example","contact":{"name":"Alexis Chaname","email":"alexischaname20@gmail.com"},"version":"Version 1"},"servers":[{"url":"http://localhost:8080","description":"Spring.boot-api-example"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Create a new Player","description":"Return the player created with your id. If a player exists with the same userName or email, then return 404 Additionally, the email must be valid and the password must have at least 8 characters and contain at least one number one lower letter, one uppercase letter and one special character.","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"400":{"description":"userName or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a player by id","description":"Return a player by the id. If the player doesn't exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the map","description":"If the app is alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","username"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"The player id","format":"int64","example":1},"username":{"title":"Player user name","type":"string","description":"The player user name","example":"myUserName"},"password":{"title":"Player password","type":"string","description":"The player password","example":"Password03#"},"email":{"title":"Player email","type":"string","description":"The player email","example":"email@gmail.com"},"avatar":{"title":"Player avatar url","type":"string","description":"The player avatar","nullable":true,"example":"https://localhost:8080/avatars/myUserName"},"last_login":{"title":"Player last login","type":"string","description":"The player last login","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}